별점 서비스 기본 API 작성, Swagger Enabled

+++ src/main/java/com/skcc/tes/starrating/domain/StarRateRepository.java
@@ -2,6 +2,6 @@ package com.skcc.tes.starrating.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StarRateRepository extends JpaRepository<StarRate, Integer>{
+public interface StarRateRepository extends JpaRepository<StarRate, Long>{
     
 }
